Allow PR status to be updated on edits
Until now, editing a PR did not update the status of
the infrastructure specs, which was annoying because often
they get fixed independently of the PR, but the PR used to
stay invalidated until a new commit was performed.

From now on, adding or editing a comment on the PR will
trigger an update of the infrastructure specs status.